move GPIO struct from peripherals to gpio module
