+++ chips/cc2650/src/peripheral_registers.rs
@@ -4,22 +4,6 @@ pub const IOC_BASE: usize = 0x4008_1000;
 
 pub const GPIO_BASE: usize = 0x4002_2000;
 
-#[repr(C)]
-pub struct GPIO {
-    _reserved0: [u8; 0x90],
-    pub dout_set: VolatileCell<u32>,
-    _reserved1: [u8; 0xC],
-    pub dout_clr: VolatileCell<u32>,
-    _reserved2: [u8; 0xC],
-    pub dout_tgl: VolatileCell<u32>,
-    _reserved3: [u8; 0xC],
-    pub din: VolatileCell<u32>,
-    _reserved4: [u8; 0xC],
-    pub doe: VolatileCell<u32>,
-    _reserved5: [u8; 0xC],
-    pub evflags: VolatileCell<u32>,
-}
-
 pub const PRCM_BASE: usize = 0x4008_2000;
 
 #[repr(C)]
